fix(gui): change order of introduction slides, to asking for fiat price preference at the end
