+++ src-gui/src/renderer/components/modal/introduction/slides/Slide07_ReachOut.tsx
